create model and modify UserRepository
Model created for PlayingUser and User, the UserRepository is
used in the domain to query the repository and to perform
CRUD operations in general.
configurations are handled using 'config' package, that allows
hierarchical configurations ('test' when testing with jest).
stub methods created to replace the repository implementation
inside the model