add methods to UserRepository and changes on the model/db mapping
every db entity now contains the methods to be cast into a
domain entity, with the respective way to be updated from it.

Note: the model might be partially incorrect because even after
updating foreign key using mixin methods, the property to access
the associated entity is not updated. this might be a bug, given
the recursive relationships.